Wipe out old messages for devices that may reuse old ids.
Deviceids are recycled like file descriptors in unix.  However the
message database is racey during removal and we often end up with
messages in the DB despite attempts to clean it up.  To workaround
this we add another db clear of the message database for the device id
being used during the device-add phase.